balance need to be fixed

+++ src/main/java/com/vastika/bank_app/dao/AccountBalanceDao.java
@@ -10,7 +10,9 @@ import com.vastika.bank_app.model.AccountBalance;
  **/
 public interface AccountBalanceDao {
 
-	int depositeAmount(AccountBalance balance);
+	double depositeAmount(AccountBalance balance);
 	
-	int withdrawAmount(AccountBalance balance); 
+	double withdrawAmount(AccountBalance balance); 
+	
+	
 }
