+++ src/main/java/com/vastika/bank_app/service/AccountBalanceService.java
@@ -11,7 +11,7 @@ import com.vastika.bank_app.model.AccountBalance;
 public interface AccountBalanceService {
 
 	
-	int depositeAmount(AccountBalance balance);
+	double depositeAmount(AccountBalance balance);
 	
-	int withdrawAmount(AccountBalance balance); 
+	double withdrawAmount(AccountBalance balance); 
 }
